containers/mocked-raw/requirements.txt
azure-cli==2.49.0
azure-cli-core==2.49.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
polars==0.19.0
pyarrow==13.0.0